Extract `ThreadSafeUnorderedMap` into a template file
